refactor: rename Article.vue -> TheArticle.vue
