+++ src/main.ts
@@ -2,7 +2,7 @@ import { ViteSSG } from "vite-ssg";
 import generatedRoutes from "virtual:generated-pages";
 import { setupLayouts } from "virtual:generated-layouts";
 import Center from "@/components/Center.vue";
-import Article from "@/components/Article.vue";
+import TheArticle from "@/components/TheArticle.vue";
 import DateText from "@/components/DateText.vue";
 import VueGtag from "vue-gtag";
 import App from "./App.vue";
@@ -16,7 +16,7 @@ export const createApp = ViteSSG(
   { routes, base: import.meta.env.BASE_URL },
   ({ app, router }) => {
     app.component("Center", Center);
-    app.component("Article", Article);
+    app.component("TheArticle", TheArticle);
     app.component("LastUpdatedAt", DateText);
     if (
       import.meta.env.MODE === "production" &&
