+++ vite.config.ts
@@ -31,7 +31,7 @@ export default defineConfig({
 
     // https://github.com/antfu/vite-plugin-md
     Markdown({
-      wrapperComponent: "Article",
+      wrapperComponent: "TheArticle",
       wrapperClasses: "markdown-wrapper",
       headEnabled: true,
       markdownItSetup(md) {
